// src/app/pages/models/informacion-usuario.model.ts
// Generated by https://quicktype.io

export interface InformacionUsuarioModel {
  id:                   number;
  permitir:             boolean;
  direccionIp:          string;
  registrosDescargados: number;
  nombre:               string;
  correoElectronico:    string;
  celular:              string;
  descargasRestantes:   number;
  mensaje:              string;
}